Improve agent RPC request error handling (#2008)
* Raise server /node_service_pods/list errors to agent

* raise server rpc handler errors

* agent: simplify manager requests

* fixup agent specs

* fixup server specs

* rescue errors in managers

* specfix

* agent RpcClient#request_with_error with RpcHelper#rpc_request wrapper

* keep a RpcClient#request compat wrapper that just logs a warning and returns nil on errors

* update rpc_request callers

* tweak RpcClient request waits

* spec RpcClient#request

* mergefix agent rpc_client specs

* spec rpc_client request_id collisions

* specfix for WaitHelper warning

+++ agent/lib/kontena/rpc_client.rb
@@ -1,9 +1,14 @@
 require_relative 'logging'
+require_relative 'helpers/wait_helper'
+
 
 module Kontena
   class RpcClient
     include Celluloid
     include Kontena::Logging
+    include Kontena::Helpers::WaitHelper
+
+    REQUEST_ID_RANGE = 1..2**31
 
     class Error < StandardError
       attr_reader :code
@@ -33,26 +38,53 @@ module Kontena
       logger.error exc.message
     end
 
+    # This method should not raise, or the Actor will crash, and terminate any other pending requests.
+    #
     # @param [String] method
     # @param [Array] params
-    # @return [Object]
-    def request(method, params)
+    # @param [Fixnum] timeout seconds
+    # @return [Object, Exception]
+    def request_with_error(method, params, timeout: 30)
       id = request_id
       @requests[id] = nil
-      sleep 0.01 until @client.connected?
+
+      if !wait_until("websocket client is connected", timeout: timeout, interval: 0.1) { @client.connected? }
+        return nil, TimeoutError.new(500, 'WebsocketClient is not connected')
+      end
+
       @client.send_request(id, method, params)
-      time = Time.now.utc
-      until !@requests[id].nil?
-        sleep 0.01
-        raise TimeoutError.new(500, 'Request timed out') if time < (Time.now.utc - 30)
+
+      if !wait_until("request #{method} has response wth id=#{id}", timeout: timeout, interval: 0.01) { @requests[id] }
+        return nil, TimeoutError.new(500, 'Request timed out')
       end
+
       result, error = @requests.delete(id)
+
       if error
-        raise Error.new(error['code'], error['message'])
+        return result, Error.new(error['code'], error['message'])
+      else
+        return result, nil
       end
-      result
     end
 
+    # Async request wrapper.
+    #
+    # Logs a warning and returns nil on errors.
+    # Use Kontena::Helpers::RpcError.rpc_request to get a raised error instead.
+    #
+    # @return [Object, nil]
+    def request(method, params, **opts)
+      result, error = request_with_error(method, params, **opts)
+
+      if error
+        warn "RPC request #{method} failed: #{error}"
+        return nil
+      else
+        return result
+      end
+    end
+
+    # Called from Kontena::WebsocketClient in the EM thread
     def handle_response(response)
       type, msgid, error, result = response
       @requests[msgid] = [result, error]
@@ -60,10 +92,13 @@ module Kontena
 
     # @return [Fixnum]
     def request_id
-      id = -1
-      until id != -1 && !@requests[id]
-        id = rand(2_147_483_647)
+      id = rand(REQUEST_ID_RANGE)
+
+      while @requests.has_key?(id)
+        sleep 0.001
+        id = rand(REQUEST_ID_RANGE)
       end
+
       id
     end
   end
